services/web/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
itsdangerous==2.0.1
Flask-GraphQL
graphene-sqlalchemy>=2.0
graphene_sqlalchemy_filter
faker
pandas
openpyxl